backend/requirements.txt
sentence-transformers==2.2.2
huggingface_hub==0.20.3
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_experimental
langchain-google-genai
pinecone
python-dotenv
Flask-Session
markdown
sentence-transformers

